Change - eliminate most global resource defaults
This huge commit changes almost every global resource default into
a per-resource default to drastically limit the scope to which these
apply.  Puppet should have offered this from the default but alas it's
relatively new.  This is a much more sane approach and has already shown
several bugs being resolved which aren't going to be detailed here --
the analysis for that would be just too costly, which is another
demonstration of just how fucked up the original Puppet design was.